Renames, SDK updates, normalize errors.
