+++ azurerm/internal/services/network/route_filter_data_source.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2019-09-01/network"
+	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2020-05-01/network"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
@@ -80,9 +80,9 @@ func dataSourceArmRouteFilterRead(d *schema.ResourceData, meta interface{}) erro
 	resp, err := client.Get(ctx, resourceGroup, name, "")
 	if err != nil {
 		if utils.ResponseWasNotFound(resp.Response) {
-			return fmt.Errorf("Error: Route Filter %q (Resource Group %q) was not found", name, resourceGroup)
+			return fmt.Errorf("Route Filter %q (Resource Group %q) was not found", name, resourceGroup)
 		}
-		return fmt.Errorf("Error making Read request on Azure Route Filter %q: %+v", name, err)
+		return fmt.Errorf("making Read request on Azure Route Filter %q: %+v", name, err)
 	}
 
 	d.SetId(*resp.ID)
